pyproject.toml
[build-system]
requires = ["setuptools~=80.9"]
build-backend = "setuptools.build_meta"

[project]
name = "zubax-bro"
authors = [
    {name = "Pavel Kirienko", email = "pavel.kirienko@zubax.com"},
]
readme = "README.md"
requires-python = ">=3.10"
dynamic = ["version"]
dependencies = [
    "openai~=1.101",
    "PyAutoGUI~=0.9",
    "pynput~=1.8",
    "mss~=10.1",
    "Pillow~=11.1",
    "tenacity~=9.1",
    "prompt-toolkit~=3.0",
    "colorlog~=6.9",
    "flask~=3.1",
]

[project.scripts]
bro = "bro.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.dynamic]
version = { attr = "bro.__version__" }

# --------------------------------------------------  BLACK  --------------------------------------------------
[tool.black]
line-length = 120
target-version = ['py313']
include = '''
(src/bro/.+\.py$)
|
(noxfile\.py$)
'''
